fix(sim): switch SimStep i128 fields to i64 to fix scenario parsing
MintTokens/BurnTokens/ConvictionCreateProposal/ConvictionStake all
carried an i128 field directly inside SimStep, a #[serde(tag = "type")]
internally-tagged enum. Serde buffers such enum variants through
serde::private::de::Content before re-dispatching into the target
type, and Content has no i128/u128 representation (serde-rs/serde#1183)
— so any i128 field inside SimStep fails to parse with "i128 is not
supported", regardless of the JSON value's magnitude. This was latent
since MintTokens/BurnTokens (and now the Conviction* steps) were never
exercised by a shipped scenario until conviction_voting.json (#1153),
which is what surfaced it: `cargo run -- run-all` — what
.github/workflows/simulation.yml runs — calls load_scenario(), which
process::exit(1)s on any parse failure, so this one bad file was
aborting the whole run and silently skipping every scenario after it
in whatever order std::fs::read_dir returned (including, in some
orderings, tools/sim/src/scenarios/optimistic_governance.json from the
previous commit).

Switch all four fields to i64 (comfortably wide enough for simulation
amounts) and widen to i128 at the three call sites that need the real
contract-facing type. No scenario JSON needs to change since existing
i128 literals were always well within i64 range.

Verified `cargo run -- run-all` now reaches and passes all 17 shipped
scenarios (conviction_voting: 15/15, optimistic_governance: 16/16),
`python3 scripts/check_budget_warnings.py` reports no warnings for any
of them, `cargo test` is 10/10 (previously 9/10 —
test_all_shipped_scenarios_are_structurally_valid now passes), and
`cargo clippy --all-targets` is clean.

+++ tools/sim/src/scenario.rs
@@ -106,11 +106,19 @@ pub enum SimStep {
     },
     MintTokens {
         actor: String,
-        amount: i128,
+        // i64, not i128: this field lives inside a `#[serde(tag = "type")]`
+        // internally-tagged enum, and serde buffers such variants through
+        // `serde::private::de::Content` before re-deserializing into the
+        // target type — a buffer that has no i128/u128 representation, so
+        // any i128 field here fails to parse with "i128 is not supported"
+        // regardless of the actual value (see serde-rs/serde#1183). Scenario
+        // token amounts comfortably fit i64; widened to i128 at the call
+        // site where the real i128 contract argument is needed.
+        amount: i64,
     },
     BurnTokens {
         actor: String,
-        amount: i128,
+        amount: i64,
     },
     UpdateConfig {
         actor: String,
@@ -179,12 +187,13 @@ pub enum SimStep {
         target: String,
         fn_name: String,
         calldata: Option<String>,
-        requested_amount: i128,
+        // i64, not i128 — see the comment on `MintTokens::amount` above.
+        requested_amount: i64,
     },
     ConvictionStake {
         actor: String,
         proposal_id: u64,
-        amount: i128,
+        amount: i64,
     },
     ConvictionWithdrawStake {
         actor: String,
